be/requirements.txt
amqp==5.1.1
anyio==3.6.2
apache-libcloud==3.4.1
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
bcrypt==4.0.1
billiard==3.6.4.0
black==22.12.0
cachelib==0.9.0
cachetools==5.2.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==38.0.4
decorator==5.1.1
dogpile.cache==1.1.8
fastapi==0.89.1
flake8==6.0.0
Flask==2.2.2
Flask-Session==0.4.0
google-auth==2.15.0
h11==0.14.0
idna==3.3
importlib-metadata==5.2.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
keystoneauth1==5.1.0
kombu==5.2.4
kubernetes==25.3.0
MarkupSafe==2.1.1
mccabe==0.7.0
munch==2.5.0
mypy-extensions==0.4.3
netifaces==0.11.0
oauthlib==3.2.2
openstacksdk==0.103.0
os-service-types==1.7.0
paramiko==2.12.0
pathspec==0.10.3
pbr==5.11.0
platformdirs==2.6.2
prompt-toolkit==3.0.37
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pyflakes==3.0.1
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.27.1
requests-oauthlib==1.3.1
requestsexceptions==1.4.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
stevedore==4.1.1
timeout-decorator==0.5.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.8
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.4.2
Werkzeug==2.2.2
zipp==3.11.0
prometheus-client==0.16.0
paramiko